Added strings for display in player text views.
initialized a textView currentPlayer.
Added method (setCurentPlayer(char) )and instance variable (mark) for setting the current player as well as changing the text displayed in the currentPlayer textView.
Added method setWinner(char) for setting the currentPlayer TextView to a win message.
Added a method placeMarkOnBoard(int row, int col) for placing a mark on the board

Fixed:
Buttons 2-9 setting the listener under them. They also need to be final if they are not instanced variables. They should be instance variables.

Ext:
Added an example that is commented out called buttonEXMP. This can be used as a template to create other buttons.